requirements.txt
fastapi
uvicorn
transformers
torch
spacy
pymupdf
python-docx
pillow
pytesseract
en-core-web-sm